+++ contracts/periphery/libraries/NFTSVG.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.7.6;
 
 import "contracts/core/libraries/TickMath.sol";
 import "@openzeppelin/contracts/utils/Strings.sol";
-import "@openzeppelin/contracts/math/SafeMath.sol";
+import "@openzeppelin/contracts/utils/math/SafeMath.sol";
 import "base64-sol/base64.sol";
 
 /// @title NFTSVG
@@ -187,6 +187,6 @@ library NFTSVG {
             tick = tick * -1;
             sign = "-";
         }
-        return string(abi.encodePacked(sign, uint256(tick).toString()));
+        return string(abi.encodePacked(sign, uint256(int256(tick)).toString()));
     }
 }
